+++ exonum-java-binding-core/rust/integration_tests/tests/utils_errors.rs
@@ -33,7 +33,8 @@ fn panic_on_exception_catch_exact_class() {
         .with_attached(|env: &JNIEnv| {
             panic_on_exception(env, throw(env, EXCEPTION_CLASS));
             Ok(())
-        }).unwrap();
+        })
+        .unwrap();
 }
 
 #[test]
@@ -43,7 +44,8 @@ fn panic_on_exception_catch_subclass() {
         .with_attached(|env: &JNIEnv| {
             panic_on_exception(env, throw(env, ARITHMETIC_EXCEPTION_CLASS));
             Ok(())
-        }).unwrap();
+        })
+        .unwrap();
 }
 
 #[test]
@@ -53,7 +55,8 @@ fn panic_on_exception_catch_jni_error() {
         .with_attached(|env: &JNIEnv| {
             panic_on_exception(env, make_jni_error());
             Ok(())
-        }).unwrap();
+        })
+        .unwrap();
 }
 
 #[test]
@@ -62,7 +65,8 @@ fn panic_on_exception_dont_catch_good_result() {
         .with_attached(|env: &JNIEnv| {
             panic_on_exception(env, Ok(()));
             Ok(())
-        }).unwrap();
+        })
+        .unwrap();
 }
 
 #[test]
@@ -80,7 +84,8 @@ fn check_error_on_exception_catch_java_error_subclass() {
     EXECUTOR
         .with_attached(|env: &JNIEnv| {
             Ok(check_error_on_exception(env, throw(env, OOM_ERROR_CLASS)))
-        }).unwrap()
+        })
+        .unwrap()
         .unwrap();
 }
 
@@ -92,7 +97,8 @@ fn check_error_on_exception_catch_java_exception_exact_class() {
                 .map_err(|e| assert!(e.starts_with("Java exception: java.lang.Exception")))
                 .expect_err("An exception should lead to an error");
             Ok(())
-        }).unwrap();
+        })
+        .unwrap();
 }
 
 #[test]
@@ -100,11 +106,13 @@ fn check_error_on_exception_catch_java_exception_subclass() {
     EXECUTOR
         .with_attached(|env: &JNIEnv| {
             check_error_on_exception(env, throw(env, ARITHMETIC_EXCEPTION_CLASS))
-                .map_err(|e| {
-                    assert!(e.starts_with("Java exception: java.lang.ArithmeticException"))
-                }).expect_err("An exception should lead to an error");
+                .map_err(
+                    |e| assert!(e.starts_with("Java exception: java.lang.ArithmeticException")),
+                )
+                .expect_err("An exception should lead to an error");
             Ok(())
-        }).unwrap();
+        })
+        .unwrap();
 }
 
 #[test]
@@ -137,7 +145,8 @@ fn get_and_clear_java_exception_if_exception_occurred() {
             );
             assert!(!env.exception_check()?);
             Ok(())
-        }).unwrap();
+        })
+        .unwrap();
 }
 
 #[test]
@@ -147,7 +156,8 @@ fn get_exception_message_without_message() {
             let exception = env.new_object(ARITHMETIC_EXCEPTION_CLASS, "()V", &[])?;
             assert_eq!(get_exception_message(env, exception)?, None);
             Ok(())
-        }).unwrap();
+        })
+        .unwrap();
 }
 
 #[test]
@@ -162,7 +172,8 @@ fn get_exception_message_with_message() {
                 Some(CUSTOM_EXCEPTION_MESSAGE.to_string())
             );
             Ok(())
-        }).unwrap();
+        })
+        .unwrap();
 }
 
 #[test]
@@ -172,7 +183,8 @@ fn get_and_clear_java_exception_if_no_exception_occurred() {
         .with_attached(|env: &JNIEnv| {
             let _exception = get_and_clear_java_exception(env);
             Ok(())
-        }).unwrap();
+        })
+        .unwrap();
 }
 
 fn throw(env: &JNIEnv, exception_class: &str) -> JniResult<()> {