+++ exonum-java-binding-core/src/main/java/com/exonum/binding/service/adapters/UserServiceAdapter.java
@@ -33,15 +33,20 @@ import com.exonum.binding.transaction.Transaction;
 import com.exonum.binding.transport.Server;
 import com.google.inject.Inject;
 import io.vertx.ext.web.Router;
+import java.util.Arrays;
 import java.util.List;
 import java.util.OptionalInt;
 import javax.annotation.Nullable;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 /**
  * An adapter of a user-facing interface {@link Service} to an interface with a native code.
  */
 @SuppressWarnings({"unused", "WeakerAccess"})  // Methods are called from the native proxy
 public class UserServiceAdapter {
+
+  private static final Logger logger = LogManager.getLogger(UserServiceAdapter.class);
 
   private static final String API_ROOT_PATH = "/api";
 
@@ -79,16 +84,25 @@ public class UserServiceAdapter {
    * @throws IllegalArgumentException if message is not a valid transaction message of this service
    */
   public UserTransactionAdapter convertTransaction(byte[] transactionMessage) {
-    BinaryMessage message = BinaryMessage.fromBytes(transactionMessage);
-    assert message.getServiceId() == getId() :
-        "Message id is distinct from the service id";
+    try {
+      BinaryMessage message = BinaryMessage.fromBytes(transactionMessage);
+      assert message.getServiceId() == getId() :
+          "Message id is distinct from the service id";
 
-    Transaction transaction = service.convertToTransaction(message);
-    checkNotNull(transaction, "Invalid service implementation: "
-            + "Service#convertToTransaction must never return null.\n"
-            + "Throw an exception if your service does not recognize this message id (%s)",
-        message.getMessageType());
-    return new UserTransactionAdapter(transaction, viewFactory);
+      Transaction transaction = service.convertToTransaction(message);
+      checkNotNull(transaction, "Invalid service implementation: "
+              + "Service#convertToTransaction must never return null.\n"
+              + "Throw an exception if your service does not recognize this message id (%s)",
+          message.getMessageType());
+      return new UserTransactionAdapter(transaction, viewFactory);
+    } catch (NullPointerException | IllegalArgumentException e) {
+      logger.warn("Failed to convert transaction {}", Arrays.toString(transactionMessage), e);
+      throw e;
+    } catch (Exception e) {
+      logger.error("Unexpected exception occurs at convert transaction {}",
+          Arrays.toString(transactionMessage), e);
+      throw e;
+    }
   }
 
   /**
@@ -110,7 +124,11 @@ public class UserServiceAdapter {
           .map(HashCode::asBytes)
           .toArray(byte[][]::new);
     } catch (CloseFailuresException e) {
+      logger.error("Failed to close some resources at getStateHashes", e);
       throw new RuntimeException(e);
+    } catch (Exception e) {
+      logger.error("Unexpected exception occurs at getStateHashes", e);
+      throw e;
     }
   }
 
@@ -125,25 +143,35 @@ public class UserServiceAdapter {
    */
   public @Nullable String initialize(long forkHandle) {
     assert forkHandle != 0;
+
     try (Cleaner cleaner = new Cleaner("UserServiceAdapter#initialize")) {
       Fork fork = viewFactory.createFork(forkHandle, cleaner);
       return service.initialize(fork)
           .orElse(null);
     } catch (CloseFailuresException e) {
+      logger.error("Failed to close some resources at initialize", e);
       throw new RuntimeException(e);
+    } catch (Exception e) {
+      logger.error("Unexpected exception occurs at initialize", e);
+      throw e;
     }
   }
 
   public void mountPublicApiHandler(long nodeNativeHandle) {
-    checkState(node == null, "There is a node already: are you calling this method twice?");
-    node = new NodeProxy(nodeNativeHandle, viewFactory);
-    Router router = server.createRouter();
-    service.createPublicApiHandlers(node, router);
-    server.mountSubRouter(serviceApiPath(), router);
+    try {
+      checkState(node == null, "There is a node already: are you calling this method twice?");
+      node = new NodeProxy(nodeNativeHandle, viewFactory);
+      Router router = server.createRouter();
+      service.createPublicApiHandlers(node, router);
+      server.mountSubRouter(serviceApiPath(), router);
+    } catch (Exception e) {
+      logger.error("Unexpected exception occurs at mountPublicApiHandler", e);
+      throw e;
+    }
   }
 
   /**
-   * Handles block commited event. This handler is invoked after commit of the block.
+   * Handles block committed event. This handler is invoked after commit of the block.
    * @param snapshotHandle a handle to a native snapshot object
    * @param validatorId a validator id. Negative if this node is not a validator
    * @param height the current blockchain height
@@ -158,9 +186,19 @@ public class UserServiceAdapter {
           : OptionalInt.empty();
       BlockCommittedEvent event =
           BlockCommittedEventImpl.valueOf(snapshot, optionalValidatorId, height);
-      service.afterCommit(event);
+      doAfterCommit(event);
     } catch (CloseFailuresException e) {
+      logger.error("Failed to close some resources at afterCommit", e);
       throw new RuntimeException(e);
+    }
+  }
+
+  private void doAfterCommit(BlockCommittedEvent event) {
+    try {
+      service.afterCommit(event);
+    } catch (Exception e) {
+      // swallow the exception because it occurs in a user code and it should not be propagated
+      logger.warn("An exception in after commit handler of event {}", event, e);
     }
   }
 
@@ -175,8 +213,13 @@ public class UserServiceAdapter {
    * <p>Releases any resources.
    */
   public void close() {
-    if (node != null) {
-      node.close();
+    try {
+      if (node != null) {
+        node.close();
+      }
+    } catch (Exception e) {
+      logger.error("Unexpected exception occurs at close", e);
+      throw e;
     }
   }
 }