+++ exonum-java-binding-qa-service/src/test/java/com/exonum/binding/qaservice/ApiControllerIntegrationTest.java
@@ -16,15 +16,18 @@
 
 package com.exonum.binding.qaservice;
 
+import static com.exonum.binding.common.serialization.json.JsonSerializer.json;
 import static com.exonum.binding.qaservice.ApiController.BLOCKCHAIN_ALL_BLOCK_HASHES_PATH;
 import static com.exonum.binding.qaservice.ApiController.BLOCKCHAIN_BLOCK_TRANSACTIONS_PATH;
 import static com.exonum.binding.qaservice.ApiController.BLOCKCHAIN_HEIGHT_PATH;
+import static com.exonum.binding.qaservice.ApiController.GET_ACTUAL_CONFIGURATION_PATH;
 import static com.google.common.base.Preconditions.checkArgument;
 import static java.net.HttpURLConnection.HTTP_BAD_REQUEST;
 import static java.net.HttpURLConnection.HTTP_CREATED;
 import static java.net.HttpURLConnection.HTTP_INTERNAL_ERROR;
 import static java.net.HttpURLConnection.HTTP_NOT_FOUND;
 import static java.net.HttpURLConnection.HTTP_OK;
+import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.anyLong;
@@ -34,9 +37,12 @@ import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import com.exonum.binding.common.configuration.ConsensusConfiguration;
+import com.exonum.binding.common.configuration.StoredConfiguration;
+import com.exonum.binding.common.configuration.ValidatorKey;
+import com.exonum.binding.common.crypto.PublicKey;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.common.hash.Hashing;
-import com.exonum.binding.qaservice.transactions.QaTransactionGson;
 import com.exonum.binding.service.InternalServerError;
 import com.exonum.binding.service.InvalidTransactionException;
 import com.google.gson.reflect.TypeToken;
@@ -298,8 +304,7 @@ class ApiControllerIntegrationTest {
               .isEqualTo(HTTP_OK);
 
           String body = response.bodyAsString();
-          Counter actualCounter = QaTransactionGson.instance()
-              .fromJson(body, Counter.class);
+          Counter actualCounter = json().fromJson(body, Counter.class);
           assertThat(actualCounter).isEqualTo(counter);
 
           context.completeNow();
@@ -359,8 +364,7 @@ class ApiControllerIntegrationTest {
               .isEqualTo(HTTP_OK);
 
           String body = response.bodyAsString();
-          Height actualHeight = QaTransactionGson.instance()
-              .fromJson(body, Height.class);
+          Height actualHeight = json().fromJson(body, Height.class);
           assertThat(actualHeight).isEqualTo(height);
 
           context.completeNow();
@@ -392,7 +396,7 @@ class ApiControllerIntegrationTest {
               .isEqualTo(HTTP_OK);
 
           String body = response.bodyAsString();
-          Object actualHashes = QaTransactionGson.instance()
+          Object actualHashes = json()
               .fromJson(body, new TypeToken<List<HashCode>>() {
               }.getType());
           assertThat(actualHashes).isEqualTo(blockHashes);
@@ -414,13 +418,60 @@ class ApiControllerIntegrationTest {
               .isEqualTo(HTTP_OK);
 
           String body = response.bodyAsString();
-          Object actualHashes = QaTransactionGson.instance()
+          Object actualHashes = json()
               .fromJson(body, new TypeToken<List<HashCode>>() {
               }.getType());
           assertThat(actualHashes).isEqualTo(transactionHashes);
 
           context.completeNow();
         })));
+  }
+
+  @Test
+  void getActualConfiguration(VertxTestContext context) {
+    StoredConfiguration configuration = createConfiguration();
+    when(qaService.getActualConfiguration()).thenReturn(configuration);
+
+    get(GET_ACTUAL_CONFIGURATION_PATH)
+        .send(context.succeeding(response -> context.verify(() -> {
+          assertAll(
+              () -> assertThat(response.statusCode()).isEqualTo(HTTP_OK),
+              () -> {
+                String body = response.bodyAsString();
+                StoredConfiguration storedConfiguration = json()
+                    .fromJson(body, StoredConfiguration.class);
+
+                assertThat(storedConfiguration).isEqualTo(configuration);
+              });
+          context.completeNow();
+        })));
+  }
+
+  private StoredConfiguration createConfiguration() {
+    return StoredConfiguration.builder()
+        .previousCfgHash(HashCode.fromString("11"))
+        .actualFrom(1)
+        .validatorKeys(
+            singletonList(
+                ValidatorKey.builder()
+                    .consensusKey(PublicKey.fromHexString("22"))
+                    .serviceKey(PublicKey.fromHexString("33"))
+                    .build()
+            )
+        )
+        .consensusConfiguration(
+            ConsensusConfiguration.builder()
+                .roundTimeout(1)
+                .statusTimeout(2)
+                .peersTimeout(3)
+                .txsBlockLimit(4)
+                .maxMessageLen(5)
+                .minProposeTimeout(6)
+                .maxProposeTimeout(7)
+                .proposeTimeoutThreshold(8)
+                .build()
+        )
+        .build();
   }
 
   private HttpRequest<Buffer> post(String requestPath) {