+++ exonum-java-binding-qa-service/src/test/java/com/exonum/binding/qaservice/QaServiceImplIntegrationTest.java
@@ -70,6 +70,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 @ExtendWith(VertxExtension.class)
 class QaServiceImplIntegrationTest {
 
+  private static final String NO_GENESIS_BLOCK_ERROR_MESSAGE =
+      "An attempt to get the actual `height` during creating the genesis block";
+
   static {
     LibraryLoader.load();
   }
@@ -285,8 +288,9 @@ class QaServiceImplIntegrationTest {
       node = new NodeFake(db);
       setServiceNode(node);
 
-      assertThrows(RuntimeException.class, service::getHeight,
-          "An attempt to get the actual `height` during creating the genesis block");
+      Exception e = assertThrows(RuntimeException.class, () -> service.getHeight());
+
+      assertThat(e).hasMessageContaining(NO_GENESIS_BLOCK_ERROR_MESSAGE);
     }
   }
 
@@ -309,8 +313,9 @@ class QaServiceImplIntegrationTest {
       node = new NodeFake(db);
       setServiceNode(node);
 
-      List<HashCode> hashes = service.getBlockTransactions(0L);
-      assertThat(hashes).isEmpty();
+      Exception e = assertThrows(RuntimeException.class, () -> service.getBlockTransactions(0L));
+
+      assertThat(e).hasMessageContaining(NO_GENESIS_BLOCK_ERROR_MESSAGE);
     }
   }
 
@@ -334,6 +339,25 @@ class QaServiceImplIntegrationTest {
       Transaction expectedTx = new IncrementCounterTx(height, counterId);
 
       verify(node).submitTransaction(eq(expectedTx));
+    }
+  }
+
+  @Test
+  void getActualConfigurationBeforeInit() {
+    assertThrows(IllegalStateException.class,
+        () -> service.getActualConfiguration());
+  }
+
+  @Test
+  @RequiresNativeLibrary
+  void getActualConfiguration() {
+    try (MemoryDb db = MemoryDb.newInstance()) {
+      node = new NodeFake(db);
+      setServiceNode(node);
+
+      Throwable t = assertThrows(RuntimeException.class, () -> service.getActualConfiguration());
+      assertThat(t.getMessage()).contains("Couldn't not find any config for"
+          + " height 0, that means that genesis block was created incorrectly.");
     }
   }