+++ CHANGELOG.md
@@ -17,9 +17,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Support of new transaction message format added in Exonum 0.10. (#534)
-- Support of core schema API. (#548, #549)
+- Support of core schema API. (#548, #549, #571, #573)
 - Support of `Service#afterCommit(BlockCommittedEvent event)` method
   that is invoked after each block commit event. (#550)
+- Support of Json serialization in a common way. (#611)  
 
 ### Changed
 - `com.exonum.binding.storage.indices.MapEntry` moved to package